+++ doc/changelog.rst
@@ -5,9 +5,11 @@ Changelog
 
 - Allow a ringbuffer to be stopped only once the last producer has indicated
   completion, rather than the first.
-- Change :py:class:`ChunkRingStream` so that stops received from the network
+- Change :py:class:`.ChunkRingStream` so that stops received from the network
   only shut down a shared ringbuffer once all the streams have stopped. A user
   call to ``stop`` will still stop the ringbuffer immediately.
+- :py:meth:`.ChunkRingbuffer.stop` now returns a boolean to indicate whether
+  this is the first time the ringbuffer was stopped.
 
 .. rubric:: 3.5.0